chore(psr-4): finish the relocation for REST routes and the Domstream view
Two leftovers from the PSR-4 move, both invisible because the compat bridge was
absorbing them.

1. All 8 registerRestApiRoute() registrations named a legacy owa_*Controller
   class and a file under modules/<mod>/controllers/ -- a directory that has not
   existed since the relocation. Every one of those paths was dead. They worked
   only because simpleFactory() resolves the legacy name through the compat map
   and returns before it ever consults the path.

   Now registered with the PSR-4 class name and a real path. Unlike
   registerAction(), this API prefixes with $this->path rather than a hardcoded
   OWA_BASE_MODULE_DIR, so a correct path can actually be expressed here.

   This also removes core's last reliance on the compat bridge for REST routing.
   The bridge is meant to serve third-party callers of the old names, which
   Lib::factory() already documents; core resolving through it meant core could
   not be exercised bridge-free.

2. modules/Domstream/Controller/DomstreamsRestView.php was a View (it extends
   Core\View\RestApi) sitting in a Controller/ directory, in the Controller
   namespace, and was the only view in the module -- Domstream had no View/
   directory at all. Moved to modules/Domstream/View/DomstreamsRest.php,
   namespaced OWA\Module\Domstream\View, and the class suffix dropped to match
   how Base's 68 views were relocated. The compat map entry for
   owa_domstreamsRestView is retargeted, so setView('domstream.domstreamsRest')
   resolves exactly as before.

   ControllerCapabilityContractTest carried an exemption for it, its comment
   noting it was "a View misfiled into a Controller/ directory ... worth
   relocating on its own merits". The exemption is removed with the misfiling.

Coverage for the view was missing. DomStreamsRestControllerTest asserted the view
NAME the controller records -- a string -- which would pass just as happily with
the view deleted. It now also resolves that name the way displayView() does, so a
move that breaks the mapping fails in the suite instead of at runtime. Confirmed
by deliberately mis-pointing the map: the suite goes red.

That test also declared its controller path as domstream/controllers/... -- the
same removed directory. Tolerated because runControllerData() tries class_exists
before falling back to the require. Corrected to the real path.

No behaviour change. 286 tests green; verified at runtime that all 8 REST route
classes load with their declared files present (previously none of the 8 did).
