+++ modules/Domstream/View/DomstreamsRest.php
@@ -1,5 +1,5 @@
 <?php
-namespace OWA\Module\Domstream\Controller;
+namespace OWA\Module\Domstream\View;
 
 
 /**
@@ -12,7 +12,7 @@ namespace OWA\Module\Domstream\Controller;
  * View
  *
  */
-class DomstreamsRestView extends \OWA\Core\View\RestApi {
+class DomstreamsRest extends \OWA\Core\View\RestApi {
 
     function render() {
